Fix : Fixed imports after renaming interfaces and resolved the comments
