+++ src/components/RegisterForm.tsx
@@ -4,11 +4,11 @@ import { useState } from "react";
 import { useRouter } from "next/router";
 import { addAttendee, addEvent } from "@/lib/firebase/firestore";
 import { toast } from "react-toastify";
-import { RegisterFormProps, RegisterFormValues } from "@/interfaces/interfaces";
+import { RegisterFormProps, RegisterFormValues } from "@/interfaces";
 import { RegisterSchema, validationSchema } from "@/validation/EventValidation";
 import FormField from "./FormField";
 import SubmitButton from "./SubmitButton";
-import { Attendee } from "@/interfaces/interfaces";
+import { Attendee } from "@/interfaces";
 import { showToast } from "@/helpers/toast";
 import { TOAST_TYPES } from "@/constants/toastEnums";
 import { v4 as uuidv4 } from "uuid";
