+++ src/pages/add-event/index.tsx
@@ -2,7 +2,7 @@ import EventForm from "@/components/EventForm";
 import React from "react";
 
 function index() {
-  return <EventForm />;
+  return <EventForm withEdit={false} />;
 }
 
 export default index;
